added single player logic, refactored code, stats accuracy

+++ src/main/java/com/teamone/typinggame/listeners/StompDisconnectedEventListener.java
@@ -1,29 +1,48 @@
 package com.teamone.typinggame.listeners;
 
+import com.teamone.typinggame.models.Player;
+import com.teamone.typinggame.models.game.Game;
+import com.teamone.typinggame.models.game.GameInterface;
 import com.teamone.typinggame.services.game.GameServiceImpl;
+import com.teamone.typinggame.services.game.SingleGameServiceImpl;
+import com.teamone.typinggame.storage.GameStorage;
+import com.teamone.typinggame.storage.PlayerStorage;
 import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
-import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
 
-import java.util.Map;
-
 @Component
+@Slf4j
 public class StompDisconnectedEventListener implements ApplicationListener<SessionDisconnectEvent> {
 
     @Autowired
     private GameServiceImpl gameService;
+    @Autowired
+    private SingleGameServiceImpl singleGameService;
+
+    private final PlayerStorage playerStorage = PlayerStorage.getInstance();
+    private final GameStorage gameStorage = GameStorage.getInstance();
 
     @SneakyThrows
     @Override
     public void onApplicationEvent(SessionDisconnectEvent event) {
-        System.out.println(event.getMessage().getHeaders());
         if (event.getMessage().getHeaders().containsKey("simpUser")) {
             String sessionId = (String) event.getMessage().getHeaders().get("simpSessionId");
-            System.out.println("Simp session id: " + sessionId);
-            gameService.removePlayer(sessionId);
+            Player player = playerStorage.getPlayer(sessionId);
+            if (player != null) {
+                String gameId = player.getGameId();
+                GameInterface game = gameStorage.getGame(gameId);
+                if (game != null) {
+                    if (gameStorage.getGame(gameId) instanceof Game) {
+                        gameService.removePlayer(sessionId);
+                    } else {
+                        singleGameService.removePlayer(sessionId);
+                    }
+                } else log.error("onApplicationEvent -> game is null");
+            } else log.error("onApplicationEvent -> player is null");
         }
     }
 }
